+++ src/main/java/hng_java_boilerplate/video/service/VideoService.java
@@ -5,6 +5,7 @@ import hng_java_boilerplate.video.dto.VideoResponseDTO;
 import hng_java_boilerplate.video.dto.VideoStatusDTO;
 import hng_java_boilerplate.video.dto.VideoUploadDTO;
 import hng_java_boilerplate.video.entity.VideoSuite;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 
@@ -14,4 +15,5 @@ public interface VideoService {
     VideoSuite getJob(String id);
     void addUpdateRecord(String id, String filename, int progress, String currentProcess);
     DownloadableDTO downloadVideo(String jobId) throws IOException;
+    VideoResponseDTO<VideoStatusDTO> startVideoProcess(MultipartFile video, String format, String jobType) throws IOException;
 }
