+++ src/main/java/hng_java_boilerplate/video/utils/VideoUtils.java
@@ -6,7 +6,9 @@ import hng_java_boilerplate.video.dto.VideoResponseDTO;
 import hng_java_boilerplate.video.dto.VideoStatusDTO;
 import hng_java_boilerplate.video.entity.VideoSuite;
 import hng_java_boilerplate.video.exceptions.FileDoesNotExist;
+import hng_java_boilerplate.video.exceptions.JobCreationError;
 import hng_java_boilerplate.video.service.VideoService;
+import hng_java_boilerplate.video.videoEnums.VideoOutput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ByteArrayResource;
@@ -18,6 +20,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
 import java.util.UUID;
 
 
@@ -132,6 +135,13 @@ public class VideoUtils {
 
     public static String sanitizeFileName(String filename) {
         return filename.replaceAll("[^a-zA-Z0-9.-]", "_");
+    }
+
+    public static VideoOutput getMatchingFormat(String outputFormat, List<VideoOutput> audioFormatList) {
+        return audioFormatList.stream()
+                .filter(audioOutput -> outputFormat.equals(audioOutput.getStatus()))
+                .findFirst()
+                .orElseThrow(() -> new JobCreationError("output_format not supported"));
     }
 
 }
